requirements.txt
ipykernel
pandas
numpy
flask
python-dotenv
transformers
torch
torchvision
torchaudio
--extra-index-url https://download.pytorch.org/whl/cu126
scikit-learn
langchain
langchain-community
langchain-core
ollama
chromadb
faiss-cpu
together